+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from "./errors.js";
+export * from "./exceptions.js";
 export * from "./option.js";
